requirements.txt
watchdog
pillow
pyinstaller